delete files in path /public/media when use npm run fresh

+++ database/seeders/20210930144946-users.js
@@ -10,9 +10,12 @@ const directory = 'public/media'
 fs.readdir(directory, (err, files) => {
   if (err) throw err
   for (const file of files) {
-    fs.unlink(path.join(directory, file), err => {
-      if (err) throw err
-    })
+    if(file !== '.gitkeep') {
+      fs.unlink(path.join(directory, file), err => {
+        if (err) throw err
+      })
+    }
+
   }
 })
